+++ packages/create-svarta-app/src/index.ts
@@ -77,7 +77,7 @@ const managers = [
 ];
 
 (async () => {
-  console.log(chalk.grey("create-svarta-app 0.0.11"));
+  console.log(chalk.grey("create-svarta-app 0.0.12"));
 
   let manager: packageManager.Type;
   const fromCmd = process.argv[0];